// frontend/features/api/model/stationResource.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */

export interface StationResource {
  id: string;
  name: string;
  /** @nullable */
  location: string | null;
  is_active: boolean;
  /** Cast to boolean */
  organization_id: string;
  created_at: string;
  updated_at: string;
}